fix install ppgan
