--- ppgan/models/criterions/IDLoss/__init__.py
@@ -1,1 +1,0 @@
-from .id_loss import IDLoss
